@@ -1,16 +1,17 @@
 apply plugin: 'com.android.library'
 
+
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 25
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 28
+        versionCode 2
+        versionName "1.0.3"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
     }
     buildTypes {
@@ -25,20 +26,13 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    api fileTree(dir: 'libs', include: ['*.jar'])
 
-    compile 'com.android.support:appcompat-v7:25.3.1'
+    api "androidx.appcompat:appcompat:1.0.2"
 
-    compile 'com.squareup.okhttp3:okhttp:3.9.1'
-    compile 'com.parkingwang:okhttp3-loginterceptor:0.5'
+    api 'com.squareup.okhttp3:okhttp:3.9.1'
+    api 'com.annimon:stream:1.1.9'
 
-    compile 'com.parkingwang:swiss:2.8.2'
-    compile 'com.annimon:stream:1.1.9'
-
-    testCompile 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.12'
 }
-
-//// Publication config
-if (hasProperty('SONATYPE_RELEASE_URL') || hasProperty('LOCAL_RELEASE_URL')) {
-    apply from: '../maven.gradle'
-}
+apply from: '../maven.gradle'
