@@ -1,6 +1,6 @@
 package com.parkingwang.version.check;
 
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.util.Log;
 
 import com.parkingwang.version.NextVersion;
