@@ -1,7 +1,7 @@
 package com.parkingwang.version.wave;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 import com.parkingwang.version.Version;
